Adding Blog page

+++ server/src/controller/blogController.js
@@ -1,9 +1,11 @@
+import AdminAuthMiddleware from '#middlewares/AdminAuthMiddleware';
 import blogModel from '#models/blogsModel';
 import { Base } from '#utils/Base';
 import CustomError from '#utils/CustomError';
 import asyncHandler from '#utils/asyncHandler';
 import { httpStatus, httpStatusCode } from '#utils/constant';
 import { filterSort, parseFilters } from '#utils/filterSort';
+import pagination from '#utils/pagination';
 import { Router } from 'express';
 
 class BlogController extends Base {
@@ -14,11 +16,35 @@ class BlogController extends Base {
     }
 
     #initializeRoutes() {
-        this.router.post('/blogs', this.#addBlog);
+        this.router.post('/blogs', AdminAuthMiddleware, this.#addBlog);
         this.router.get('/blogs', this.#getBlog);
         this.router.get('/blogs/:id', this.#getBlogById);
-        this.router.delete('/blogs', this.#deleteBlogs);
+        this.router.delete('/blogs', AdminAuthMiddleware, this.#deleteBlogs);
+        this.router.put('/blogs', AdminAuthMiddleware, this.#updateBlog);
     }
+
+    #updateBlog = asyncHandler(async (req, res) => {
+        let { blogId, name, description, image, content, keywords } = req.body;
+        const updateBlog = await blogModel.findByIdAndUpdate(blogId, {
+            name,
+            description,
+            image,
+            content,
+            keywords,
+        });
+        if (!updateBlog) {
+            throw new CustomError(
+                'Blog does not exist.',
+                httpStatusCode.BAD_REQUEST,
+            );
+        }
+        return this.response(
+            res,
+            httpStatusCode.OK,
+            httpStatus.SUCCESS,
+            'Blog is updated successfully.',
+        );
+    });
 
     #getBlogById = asyncHandler(async (req, res) => {
         const { id } = req.params;
@@ -40,6 +66,12 @@ class BlogController extends Base {
 
     #deleteBlogs = asyncHandler(async (req, res) => {
         const { blogIds } = req.body;
+        if (req.role !== 'admin') {
+            throw new CustomError(
+                "You dont' have right to delete the blog.",
+                httpStatusCode.UNAUTHORIZED,
+            );
+        }
         const deletedData = await blogModel.deleteMany({
             _id: { $in: blogIds },
         });
@@ -68,8 +100,15 @@ class BlogController extends Base {
         page = Number(page) || 1;
         perRow = Number(perRow) || 20;
 
+        const skip = pagination(page, perRow);
+
         const [blogs, totalCount] = await Promise.all([
-            blogModel.find(filterQuery).sort(Sorts).limit(perRow).exec(),
+            blogModel
+                .find(filterQuery)
+                .sort(Sorts)
+                .skip(skip)
+                .limit(perRow)
+                .exec(),
             blogModel.countDocuments(filterQuery).exec(),
         ]);
